backend/requirements.txt
Flask
flask-cors
google-generativeai
PyPDF2
python-dotenv